[cli] add support for simple command extensions in the cli
Expo already have an existing extension point in the Expo CLI called "Dev tools plugins" (https://docs.expo.dev/debugging/devtools-plugins/), which is the ability for an Expo package (npm library) to expose a web app that can be opened and hosted by the CLI using the shift + m shortcut. A devtools plugin can communicate with running Expo apps through web sockets and there are multiple examples of plugins available, from expo-atlas to redux plugin.

A missing feature in the current devtools plugin extension API is the ability to run simple one-shot commands from the CLI using the same mechanism - but launching commands instead of a web app in the browser, but still with support communication with any running Expo app.

Use-cases for command-line based CLI-extension can be to list and test expo background tasks from the command line, query a running app for its current screen (url) in expo-router or perform tasks like checking for SDK updates through the CLI (and maybe even upgrade the current project).

I've extended the devtools configuration element in the expo-modules.config.json files so that we can define commands for the CLI using a configuration like this:

```json
{
  "platforms": [ "devtools" ],
  "devtools": {
	  "cliExtensions": {
	    "description": "Expo: tools for Expo App development",
	    "commands": [
	      { "name": "outdated", "title": "Check for outdated Expo packages", "environments": ["cli", "mcp"] },
	      { "name": "upgrade", "title": "Upgrade to latest Expo packages", "environments": ["cli", "mcp"] },
	      { "name": "take_screenshot", "title": "Take screenshot of Expo app", "environments": ["cli", "mcp"] },
	      { "name": "doctor", "title": "Run the Expo Doctor diagnostics tool", "environments": ["cli", "mcp"] }
	    ],
	    "entryPoint": "cli/build/index.js"
    }
  }
}
```

Each command will be displayed in the CLI interactive commands menu (shift+m) and will be run as node commands.

- expo-modules-autolinking:
  - Added types for the cli extension configuration
- cli:
  - Added support for showing the cli extensions as an interactive action (shift + m) in the cli
  - Added support for looking up and adding executors for cli externsions in DevToolsPluginManager.ts

  Also fixed from last try
  - Added zod validation to plugin configuration
  - Added zod validation to output
  - Added validation of parameters/arguments on execution
  - Added confirmation before running a command

  Replaces #38091